requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiogram>=3.4.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
httpx>=0.26.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
joserfc>=0.9.0
cryptography>=42.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
